loadhead modify

+++ p2p-src/Include/JEANP2PPRO.h
@@ -47,7 +47,7 @@ struct load_head {
 	char logo[4];
     u_int32_t index;
 	char priority;
-	u_int64_t length;
+	u_int32_t length;
 	unsigned char direction;//direction = 0, to master; direction = 1, to slave
 	u_int32_t totalLen;
 	u_int16_t sliceIndex;
